// src/api/model/getMangaSortBy.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Mano API
 * Mano API
 * OpenAPI spec version: 2.0
 */

export type GetMangaSortBy = typeof GetMangaSortBy[keyof typeof GetMangaSortBy];


export const GetMangaSortBy = {
  most_liked: 'most_liked',
  most_viewed: 'most_viewed',
  highest_rated: 'highest_rated',
  newest: 'newest',
  latest_chapter: 'latest_chapter',
} as const;