// src/api/model/responseMangaResponseAgeRestriction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Mano API
 * Mano API
 * OpenAPI spec version: 2.0
 */

export type ResponseMangaResponseAgeRestriction = typeof ResponseMangaResponseAgeRestriction[keyof typeof ResponseMangaResponseAgeRestriction];


export const ResponseMangaResponseAgeRestriction = {
  all: 'all',
  '12+': '12+',
  '16+': '16+',
  '18+': '18+',
} as const;